Nexus transport for workflow callers (#1)
This PR makes it possible to make MCP calls from a Temporal workflow to
a Nexus MCP server, via a custom transport for the standard MCP SDK
client. The typical use case is an AI agent implemented as a Temporal
workflow that wants to make use of tools that are themselves backed by
Temporal workflows.